Extract figure generation in separate files (#10)
This Pull Request is a proposal to organize the code for generating our figures
in a more modular way.

Instead of putting all the figures in the same `figures` file, this creates a
`figures` module, with one file per figure.

In this file, we have a class that takes `all_params` and `years` as inputs and
has one `figure()` function that generates the figure. The necessary calculations
to generate the figure are in the same class and file (instead of a separate 
`calculations` file). This makes for smaller, more self-contained files.

We also have more explicit names for our figures: `PowerGenerationCapacity` and
`PowerGenerationCapacityAggregate` instead of `fig1` and `fig2`. This will make
finding the right figures easier when we have a lot of them.

Other changes that were made to make this possible:
- Extracted the `agg_col` variable in the `app.constants` module.
- Renamed the `figures` file to `old_figures` to avoid name clash with the new
`figures` module.