Show title and description from metadata (#23)
We are now generating `metadata.json` files as part of the results file in the platform
(see https://github.com/yboulkaid/osemosys-cloud/pull/183). This PR makes it so that the visualization
reads the metadata and displays it as the title and description of the visualization.

Note: This is only going to work with runs that have been created after today (Aug. 2nd, 2020).