Fix bug where the last year is cut
We are using the `range` function to create the list of
years from a minimum and a maximum value.

The maximum value of `range` is excluded in the generated
range, which causes the last year to be truncated in some figures.

This fixes the issue by including the last year in the range.